+++ src/client/index.ts
@@ -1,4 +1,3 @@
-import { Block, Router } from './packages';
 import { render } from './utils';
 import {
 	chat,
@@ -9,8 +8,11 @@ import {
 	profile,
 	defaultPage
 } from './pages';
+// @ts-ignore
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import style from './styles/style.scss';
+import { Block } from './packages/Block';
+import { Router } from './packages/Router';
 
 const router = new Router();
