+++ src/client/packages/Router/components/index.ts
@@ -1,2 +1,1 @@
-/* eslint-disable import/prefer-default-export */
 export { default as RouterLink } from './link';
