+++ src/client/ui/TextField/index.ts
@@ -1,8 +1,8 @@
 import tmp from 'bundle-text:./index.hbs';
 import * as style from './style.module.scss';
 import { BaseInputHandlers, classNames } from '../../utils';
-import { Block, BlockEvents } from '../../packages';
 import BaseInput from '../baseInput';
+import { BlockEvents, Block } from '../../packages/Block';
 
 export type TextFieldProps = {
 	type?: HTMLInputElement['type'];
