+++ src/client/ui/container/index.ts
@@ -1,6 +1,6 @@
 import tmp from 'bundle-text:./index.hbs';
+import { Block } from '../../packages/Block';
 import * as style from './style.module.scss';
-import { Block } from '../../packages';
 
 class Container extends Block {
 	constructor(args: { children: Block | string }) {
