+++ src/client/utils/classNames.ts
@@ -1,13 +1,12 @@
 /**
- * 
  * @param  {...(string | { string: any } | any[])} args
  * @returns 
  */
 type ObjectType = Record<string, unknown>
 
-type ArrayType = (string | ObjectType | number | ArrayType)[]
+type ArrayType = (string | ObjectType | number | ArrayType | null | false | undefined)[]
 
-const classNames = (...args: ArrayType) => {
+export const classNames = (...args: ArrayType) => {
 	const calcObj = (obj: ObjectType): string => Object
 		.entries(obj)
 		.filter((entry) => !!entry[1])
@@ -18,12 +17,10 @@ const classNames = (...args: ArrayType) => {
 		.filter((item) => !!item)
 		.map((item) => {
 			if (typeof item !== 'object') return String(item).trim();
-			if (typeof item === 'object' && !Array.isArray(item)) return calcObj(item);
+			if (typeof item === 'object' && !Array.isArray(item)) return calcObj(item as ObjectType);
 			if (Array.isArray(item)) return calcArr(item);
 			throw new Error('wrong type of args');
 		}).join(' ');
 
 	return calcArr(args).trim();
 };
-
-export default classNames;
