+++ src/client/utils/index.ts
@@ -1,6 +1,9 @@
-export { default as classNames } from './classnames/index';
+export { classNames } from './classNames';
 export { default as defineMixin } from './mixin';
 export * from './validation';
 export { default as render } from './render';
 export type { default as baseAPI } from './baseRequests';
+export { set } from './set';
+export { isEqual } from './isEqual';
+export { deepClone } from './deepClone';
 export * from './functions';
