+++ src/client/utils/render.ts
@@ -1,6 +1,9 @@
-import { Block } from '../packages';
+interface IBLock {
+	getContent(): HTMLElement | null;
+	dispatchComponentDidMount(): void;
+}
 
-export default (selector: string, component: Block) => {
+export default (selector: string, component: IBLock) => {
 	const rootEl = document.querySelector(selector);
 	if (!rootEl) throw new Error('root element not found');
 	const content = component.getContent();
